Doc upd

+++ src/include/lock/spinlock.h
@@ -1,7 +1,9 @@
 #pragma once
 #define _SPINLOCK_H_
 
-typedef unsigned int spinlock_t;
+#include <stdint.h>
+
+typedef size_t spinlock_t;
 
 void spinlock_acquire(spinlock_t *);
 void spinlock_release(spinlock_t *);
